+++ AIMSICD/src/main/java/com/secupwn/aimsicd/constants/DrawerMenu.java
@@ -12,7 +12,7 @@ public class DrawerMenu {
 
     /**
      * Constants of id for NavDrawerItem
-     * <p>Relates to {@link com.secupwn.aimsicd.drawer.NavDrawerItem#getId()}<br />
+     * <p>Relates to {@link com.secupwn.aimsicd.ui.drawer.NavDrawerItem#getId()}<br />
      */
     public static class ID {
