+++ AIMSICD/src/main/java/com/secupwn/aimsicd/service/CellTracker.java
@@ -27,7 +27,7 @@ import android.telephony.TelephonyManager;
 import android.telephony.cdma.CdmaCellLocation;
 import android.telephony.gsm.GsmCellLocation;
 
-import com.secupwn.aimsicd.MainActivity;
+import com.secupwn.aimsicd.ui.activities.MainActivity;
 import com.secupwn.aimsicd.AndroidIMSICatcherDetector;
 import com.secupwn.aimsicd.BuildConfig;
 import com.secupwn.aimsicd.R;
