+++ AIMSICD/src/main/java/com/secupwn/aimsicd/ui/activities/AboutActivity.java
@@ -3,7 +3,7 @@
  * LICENSE:  http://git.io/vki47 | TERMS:  http://git.io/vki4o
  * -----------------------------------------------------------
  */
-package com.secupwn.aimsicd.activities;
+package com.secupwn.aimsicd.ui.activities;
 
 import android.content.Context;
 import android.content.Intent;
