+++ AIMSICD/src/main/java/com/secupwn/aimsicd/ui/activities/PrefActivity.java
@@ -3,16 +3,16 @@
  * LICENSE:  http://git.io/vki47 | TERMS:  http://git.io/vki4o
  * -----------------------------------------------------------
  */
-package com.secupwn.aimsicd.activities;
+package com.secupwn.aimsicd.ui.activities;
 
 import android.app.FragmentManager;
 import android.app.FragmentTransaction;
 import android.os.Bundle;
 
-import com.secupwn.aimsicd.fragments.SettingsFragment;
+import com.secupwn.aimsicd.ui.fragments.PrefFragment;
 
 
-public class SettingsActivity extends BaseActivity {
+public class PrefActivity extends BaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -21,7 +21,7 @@ public class SettingsActivity extends BaseActivity {
     }
 
     private void loadFragment() {
-        SettingsFragment settingsFragment = new SettingsFragment();
+        PrefFragment settingsFragment = new PrefFragment();
         FragmentManager fragmentManager = getFragmentManager();
         FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
         fragmentTransaction.replace(android.R.id.content, settingsFragment);
