+++ AIMSICD/src/main/java/com/secupwn/aimsicd/ui/activities/SettingsActivity.java
@@ -3,24 +3,28 @@
  * LICENSE:  http://git.io/vki47 | TERMS:  http://git.io/vki4o
  * -----------------------------------------------------------
  */
-package com.secupwn.aimsicd.activities;
+package com.secupwn.aimsicd.ui.activities;
 
 import android.app.FragmentManager;
 import android.app.FragmentTransaction;
 import android.os.Bundle;
 
-import com.secupwn.aimsicd.fragments.MapPrefFragment;
+import com.secupwn.aimsicd.ui.fragments.SettingsFragment;
 
-public class MapPrefActivity extends BaseActivity {
+
+public class SettingsActivity extends BaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        MapPrefFragment settingsFragment = new MapPrefFragment();
+        loadFragment();
+    }
+
+    private void loadFragment() {
+        SettingsFragment settingsFragment = new SettingsFragment();
         FragmentManager fragmentManager = getFragmentManager();
         FragmentTransaction fragmentTransaction = fragmentManager.beginTransaction();
         fragmentTransaction.replace(android.R.id.content, settingsFragment);
         fragmentTransaction.commit();
     }
-
 }
