+++ AIMSICD/src/main/java/com/secupwn/aimsicd/ui/fragments/CellInfoFragment.java
@@ -16,7 +16,7 @@
  * ""
  */
 
-package com.secupwn.aimsicd.fragments;
+package com.secupwn.aimsicd.ui.fragments;
 
 import android.content.ComponentName;
 import android.content.Context;
