+++ AIMSICD/src/main/java/com/secupwn/aimsicd/ui/fragments/MapPrefFragment.java
@@ -3,7 +3,7 @@
  * LICENSE:  http://git.io/vki47 | TERMS:  http://git.io/vki4o
  * -----------------------------------------------------------
  */
-package com.secupwn.aimsicd.fragments;
+package com.secupwn.aimsicd.ui.fragments;
 
 import android.os.Bundle;
 import android.preference.PreferenceFragment;
