+++ AIMSICD/src/main/java/com/secupwn/aimsicd/ui/widget/HighlightTextView.java
@@ -1,4 +1,4 @@
-package com.secupwn.aimsicd.widget;
+package com.secupwn.aimsicd.ui.widget;
 
 import android.content.Context;
 import android.util.AttributeSet;
