+++ AIMSICD/src/main/java/com/secupwn/aimsicd/utils/RequestTask.java
@@ -14,7 +14,7 @@ import android.support.v4.content.LocalBroadcastManager;
 
 import com.secupwn.aimsicd.BuildConfig;
 import com.secupwn.aimsicd.R;
-import com.secupwn.aimsicd.fragments.MapFragment;
+import com.secupwn.aimsicd.ui.fragments.MapFragment;
 import com.secupwn.aimsicd.adapters.AIMSICDDbAdapter;
 import com.secupwn.aimsicd.constants.DrawerMenu;
 import com.secupwn.aimsicd.constants.TinyDbKeys;
